Add BikeClusterView
* add BikeClusterAnnotationView
* move FriendAnnotationView to separate file

change priority to display cluster under user position

rename function

move clusterView to its own file
